Hotjar Survey + Localhost SSH Testing
The Hotjar Survey is in place, but there is an issue, we need to import it to the GitHub site in order to test it's extension. Therefore, in this export I've used a localhost SSH test (Yes, it requires HTTPS connection in order to implement the extension), but it doesn't work. We'll need to do it on the real site and complete survey also, on the Hotjar site (the persona survey).